[INTERNAL] table.SelectionPlugin: use Element instead of ManagedObject
As discussed with Core colleagues:
This seems to be the recommended/safer options, even though one may not
need all the APIs (layoutData, domRef, tooltips etc.).
Also the current UI5 tooling/infrastructure seems to expect this, so we
switch to be compatible and also avoid any other side-effects.

This also solves the issue with RTA scenario crashing!

Change-Id: I1ead07fab8619db6a50d714b17010f292cb3313b