[FIX] sap.f.DynamicPage: Space key is now allowed in Input in the title
Problem: When an Input control is placed in the title of the
DynamicPage, Space key is not allowed, because the title has logic to
prevent the default behaviour of the event.

Solution: Now the default behaviour is prevented only when the source of
the event is the title itself (not a control, placed in it).

Change-Id: Ie1251ef59b163960dc8464cc285f335fb8001304
CR-Id: 002075125900001524942019
BCP: 1980188961

(cherry picked from commit 7e460fedb44e18aaaa91621c74f78205a287b1b1)

Change-Id: If6644b529ee11ae9a00b3779193063be11b8cb31