[INTERNAL] sap.f.GridContainer: Prevent grid blowout
Previously, if an item had more columns than the total columns in the grid,
it used to break the whole layout.
Now this is prevented by limiting the columns span to the total number of columns in the grid.
Added event layoutChange, so that the developer can also handle,
when columns are reduced.

Change-Id: Iecf546ed85d2372f432b8936591d84f23d914c51
JIRA: BGSOFUIRODOPI-2014