[INTERNAL] sap.uxap.ObjectPageHeader: No enforcement for default actions
Before: We were forcing the button type of actions with "Default"
button type, to be "Transparent".

Now: By the latest requirement for Fiori 3, we are not doing it anymore.

Additional information: To keep the look of the "Default" actions in
belize theme, we are over-styling the "Default" actions to look like
"Transparent", but we are not changing their type to "Transparent".

Change-Id: I4ea823999a1bc5825ba63cdf304e412aeaeb5fd0
BCP: 1980152405