[FIX] ManagedObject: Fix regression in aggregation refresh
Note:
From commit I96a744f439a7f022cde4776d2466ca8dfafc404c we refactored the
accessors for the aggregations in the Managed Object and the binding
change handles. Before this refactoring when there was neither an
instance refresh nor update the generic update method was called in the
refresh aggregation. After our change a refresh is called which causes
the problem in the incident below.
Using a refresh instead of update here seemed to be more correct as the
error was inside the coding before. Now unfortunately we can not change
this any more.

Change-Id: Ida2cfe7c0e163f8f46993eec6c24f0d14acbb4bf
BCP: 1980199077