[INTERNAL][FIX] Table: Prevent focus steal by table after rendering rows
Fixes an issue introduced with Ie8ce773e6bad4791057e5c66892381b541d25613

The table would put the focus in itself after rendering the rows, even
if the focus was on an element outside the table before rendering. As
the issue depends on the row rendering mechanic, the issue would only
occur in VisibleRowCountMode=Auto on resize.

Change-Id: Idb603084ccf390df083647fbe641516b3388cb4a
BCP: 1980202585