[INTERNAL][FIX] sap.m.Select: Removed dependency to sap.ui.layout library
Problem:
The samples for sap.m.Select are not working if the
sap.ui.layout library is not loaded.

Root cause:
The sap.m.Select depends on the sap.ui.layout.SimpleFixFlex
control in order to render the value state message text
and the sap.m.SelectList. The sap.ui.layout.SimpleFixFlex
control is located in the sap.ui.layout library and thus
introduces a dependency for the sap.m library.

Solution:
Move the sap.ui.layout.SimpleFixFlex control to the sap.m
library as it is a private control that we have created.

Change-Id: I1acaa23b75c094fb99d07612044a3d5575923bb0
BCP: 1980194713