[FIX] sap.uxap.ObjectPageSubSection: Corrected sizing in Fiori3
Background:
When a subSection has the "sapUxAPObjectPageSubSectionFitContainer"
CSS class, the page internally sets its height to be the height
of the sections container (in order to ensure the subSection will
expand to exactly fit its container)

Problem:
In Fiori3, the subSection also has paddings (top and bottom) that
affect the final subSection height.
=> the final height is bigger

Solution:
Specify "border-box" value for "box-sizing" of subSections, to
ensure its height *includes* its paddings.

Change-Id: I27de6dbb390d7531cda204aed2b17f72c584a459
BCP: 1980202708