[INTERNAL] ui5loader-autoconfig: adapt shim for sap.viz library
Background:
The sap.viz charting library (not part of OpenUI5) contains two
conflicting versions of the same framework. The more modern one consists
of the 'sap-viz-info-framework' and 'sap-viz-info-charts' modules,
whereas the older one is contained in a larger bundle called 'sap-viz'.

Both frameworks contain moderator code to manage coexistence. E.g. when
the UI5 language is changed, the change listener of sap.viz calls
methods of both frameworks to update the internal language.

Unfortunately, this moderator code only works when the modern framework
has been executed first and the older one second. If the startup order
is reversed, the moderator code fails.

Normally, this doesn't matter as the sap.viz library, when loaded, only
starts the first framework. But when a module depends on both frameworks
and when async loading is active, the execution order becomes random and
conflicts may occur. This meanwhile can happen in sap.viz tests, after
migrating them to async loading.

The only safe protection against such wrong usage is a virtual (shimmed)
dependency from the old framework to the new one. This adds a
significant extra-cost to the old framework. But in UI5 usage scenarios,
this is no drawback as the sap.viz library anyhow starts the new
framework first. The only other known use cases are HTML islands in
WebDynpro ABAP where the scripts of the old framework are loaded without
the help of UI5. Because UI5 module loading is not used, the shim is
ignored and loading behavior doesn't change.

Note: it is obviously bad design to maintain the shims for sap.viz in
the OpenUI5 layer. But as long as 'manifest first' for libraries has not
been implemented, this is the easiest way to make the shims available
early enough.

Change-Id: If906e36ea7b9b95385615719ca2d4eb4adce78bc