[FIX] mvc/View: Deregister erroneous instances from the Core
When a View is instantiated with the 'async' setting it can occur
that Errors appearing in the async part of the View creation lead
to a bypassing of the deregistration normally done by the Managed-
Object constructor.
With this change the internal Promise is appended with a generic
catch handler that deregisters the erroneous instance.

Change-Id: Icce6f3cffeeaeeab0da0349c1de71d19798e7d5a
BCP: 1980108149