[FIX] v2.ODataListBinding: Initialize relative bindings
If a binding is relative to a Context that was created via ‚createEntry‘
the binding will not be initialaized. This prevents the binding from
triggering requests to a non existing entity.
When the ‚created‘ entity was submitted successfully the binding needs
a refresh as it could work as expected now.

BCP: 1970130260

Change-Id: Ia15244c5dee53cd4beb397e0a39a40d6e61ba205