[INTERNAL] sap.m.MultiInput: Clear dialog's value after selection
The fix is applicable for mobile scenarios. Previosly, if a
suggestion item was selected from the dialog, the dialog input
was updated to match the title of the item. This behavior was
only observed on mobile.

Now the input value is cleared after selection of a suggestion
item both on desktop and on mobile.

Change-Id: I817eadf697f09bc2f33da655b45d3d242f6dafbc
JIRA: BGSOFUIRILA-2383