[FIX] BindingParser: detect complex binding that uses quoted properties
The binding parser did not correctly detect a complex binding when the
first property of the binding used a quoted property name (e.g. when
standard JSON notation was used).

With this change, any valid property syntax is recognized (unquoted
property names, strings with single or double quotes) and triggers
parsing of a complex binding (vs. simple binding).

Change-Id: Ibcd93bb17b679e97e3e050fc8ecc68b8680e6d20