[FIX] Table: Set correct firstVisibleRow also if parent is invisible
Fixes an issue where the table scrolls back to the top, because it
operates with a scrollbar element that was removed from the DOM without
notifying the table.
This can be the case, for example, if the parent of the table was made
invisible and the firstVisibleRow property of the table was changed.

Change-Id: Ied09be0bbcc4a20158f4a0538a7c3d353f1401a2
BCP: 1980220195