[FIX] Table: Set correct firstVisibleRow also if parent is invisible
Fixes an issue where the table scrolls back to the top, because it
operates with a scrollbar element that was removed from the DOM without
notifying the table.
This can be the case, for example, if the parent of the table was made
invisible and the firstVisibleRow property of the table was changed.

Change-Id: I703be10c05c0a97be86e615aa47679c3f1e78042
BCP: 1980220195
CR-Id: 002075125900003020452019