[FIX] sap.m.DatePicker: Support for data types is now improved
Problem: When formatting and parsing, the DatePicker control is using
the data types properties to instantiate sap.ui.core.format.DateFormat.
Some data types create their DateFormat instance lazily, so those
properties may not be available yet. This will result in creating a
formatter that is not matching the formatter of the provided by the
app developer data type.

Solution: Instead of trying to replicate internally in the DatePicker
the formatter that is (or is going to be) created by the data type, we
reuse the same formatter that is created by the data type.

The solution is not perfect, as we cannot support all the currently
existing data types or the future (or custom) data types that can be
created, but it is trying to provide a better support for limited set
ot data types with limited set of their formatting options and
constraints. With this particular fix we provide a better support for
sap.ui.model.odata.type.DateTime with constraint displayFormat: Date.
Supporting the default constraint displayFormat: DateTime, for example,
is not possible as it contains time part, that is not relevant for the
DatePicker control.

Change-Id: I55770d53d90322317fcbcc084181e3f022bba8af
BCP: 1870481844