[FIX] sap.m.NotificationListGroup: Items position is now read out correctly
Problem1 : 'aria-hidden' was set to 'true' and this stopped
the screen reader to read the position of the elements
(for example 1 of 4).
Solution1: Add 'aria-hidden' only when it is needed, and remove
it when it is not.

Problem2: The position of the group itself wasn't read out.
Solution2: Set 'role' to 'option'.

Change-Id: Icf20bc32bf3fbea7f04cedc3520f54f2f1f82ecc
BCP: 1980211051