[FIX] sap.m.NotificationListItemBase: Focusing parent after close is now working correctly
Problem: When we have a NotificationListItem and press the
close button, we expect that the item will be destroyed or removed.
So we add an event delegate that will focus the parent of the item.
This breaks cases where the item doesn't get destroyed or removed.

Solution: Add the delegate which changes the focus only in cases
where the item is really destroyed or removed.

Change-Id: I54e5dd30f8619298067b2b93176a3d76cf6ddb3f
BCP: 1980098133