[FEATURE] ui5loader: support UMD shim also in async mode
To support coexistence with other module loaders, especially with
require.js (which is mandatory when using the sap.viz charting library),
the ui5loader in sync mode did hide such loaders before executing a
3rdparty UMD module and restored that loader afterwards.

In async mode, this strategy wasn't applicable as loading via scripts
does not allow to act immediately before a script starts. Disabling the
loader for a longer time (from script creation until the 'load' event
for the script) could interfere with intentional calls to the other
loader.

With this change, a different mechanism has been implemented for async
mode: Access to the define.amd property is intercepted. On read, the
current module is checked (via the internal execution stack in case of
sync execution or via document.currentScript in case of async execution)
and when the module has been requested by the ui5loader AND is known to
be an UMD module, then 'undefined' is returned instead of the real value
of define.amd. That way, a 3rdparty AMD loader becomes invisible for
modules required by UI5.

For IE11, which does not support document.currentScript, the loader
falls back to sync loading for UMD modules. This approach is not CSP
compliant, but IE11's CSP support isn't good anyhow.

To deal with lazy loading of a 3rdparty AMD loader, access to
the global 'define' property is also intercepted. This is unfortunately
not possible when require.js was loaded before UI5. require.js creates
the globals 'define', 'require' and 'requirejs' with var statements.
Those are implicitly defined as 'configurable:false' and re-defining
them with Object.defineProperty fails. The loader logs a warning in that
case, but still intercepts access to define.amd.

Change-Id: Idc22ea2f0f489cc4ccbd87272162c5bd0e61937e