[INTERNAL] sap.ui.model.odata.v4.ODataModel: operations returning collection
The problem is that the property binding asks the list binding whether
it can use its cache when the execute is already finished and the list
is immutable. Before the change, the list binding added the key
properties to the aggregated $select and thus accepted key properties,
but not non-key properties.

POC1 ensures that a binding below a deferred operation binding accepts
a property binding regardless of its path.

POC2 ensures that an immutable binding does not add the key properties
to the aggregated $select and accepts a property binding if its path is
in the aggregated $select or if the aggregated $select is empty.

PS1: red integration test
PS2: further integration tests adjusted; POC1
PS3: POC2
PS4: ODataParentBinding#fetchIfChildCanUseCache always resolves with
  true if binding is operation or depends on operation: we do not
  support auto-$expand/$select yet on operations and assume the
  operation result contains data for dependent bindings
PS5: review comment
PS6: ODataParentBinding#fetchIfChildCanUseCache back to state of PS4
  => tests green
PS7: review comments

Change-Id: I396cf5a8ae7293b416dec6ba47045016bcc71ba0
BCP: 1980252607