[INTERNAL] sap.m.NavContainer: Improved slide and fade animations
Improvement in fade animations: Now the "FROM" page of the Nav Container
fades out and then the "TO" page fades in.
Improvement in slide animations: Now, when you are navigating forward
the "FROM" page slides 1rem distance on the left, while fading out,
and the "TO" page also slides 1rem on the left, while fading in. When
moving back, the two pages fades the same way, only the moving direction
is right.

Change-Id: I8b657e6e53b8c5da27e0c35626549f627567a208
JIRA: BGSOFUIPIRIN-3556