[FIX] sap.m.MultiInput,MultiComboBox: Tooltips in disabled state now shown
- The style "pointerEvents: none" is preventing all mouse events
to be executed, which means that the hover event is prevented as well.
The result is that the tooltips of the controls are not shown.
- In order to prevent the deletion of the tokens, but in the same time
keep the tooltips, we need to introduce the "enabled/disabled"
state of the controls via the EnabledPropagator.
- We also have to apply the visual specs for all themes for the Tokenizer
and the Token controls.
- For the sake of less maintainabillity we have done that in the
Tokenizer control, which is the common point where most of the styles and
logic concerning this issue is implemented.

Change-Id: I5b940884951dd5c1cd478787f81fa8a5650d1f5b
BCP: 1980173700