[FEATURE] mvc/XMLView: support core:require tag
With this change it is possible to require Modules in XMLViews and
use their static functions as event handlers, formatters or inside
of ExpressionBindings. That can help to avoid global variables or
eases reuse of certain helper classes without indirection via the
Controller.

A new private module resolveReference has been added to unify the 
coding over the three modules BindingParser, DataType and the
EventHandlerResolver.

JIRA: CPOUIFPHOENIXCORE-2660
Change-Id: I13a2fe4643b08373162184c6bffd5e086027f5e9