[INTERNAL] all: add lazy dependencies to libraries that are used in APIs
When types from another library are used in APIs of a library, a
dependency to that other library needs to be modeled (might be a lazy
dependency). Otherwise, generation of derived artifacts (e.g. TypeScript
interfaces) might fail.

Dependencies in package.json and pom.xml should include all dependencies
from .library, no matter whether they are lazy or not.


Change-Id: If30a95b4c06112d015fc90782d887ee7d8dc0be3