[INTERNAL] sap/ui/Device: Handle viewport size changes
When the viewport is set on some devices, the window's viewport size changes
as well, but the resize event handler of the Device API, which updates the Device size info,
is called too late.

With this change the Device API size info is updated directly after the meta
tag 'viewport' was added to the root HTML document.

Change-Id: I0ff512d4e02a0e169156ea4ba464c7cf83a961cc
BCP: 1980130016