[INTERNAL] InvisibleText: do not create RenderManager
 - Once the static UIArea is rendered it should not be invalidated.
Therefore when toStatic method of InvisibleText control is called it
creates new RM to render the control and append the output to the static
UIArea.
 - Creating a new RM instance during the rendering of an other control
just for a single span output seems like too much work since new RM must
create different facades and focus handler which are not needed for
InvisibleText control rendering.
 - InvisibleText#toStatic method is used by many controls for
accessibility reasons
 - This change is trying to build equivalent renderer markup and uses
DOM API to append the markup to the static UIArea. It is true that we
have now two ways to render the control but the motivation is the
simplicity of the InvisibleText control markup and the performance gain.

Change-Id: Ibc3eb68f5ccff5f5d537fa0b61cc72d867ecbb75