[FEATURE] sap.m.OverflowToolbar: Async mode for layouting is introduced
A new asyncMode property of sap.m.OverflowToolbar is introduced. By
default its value is "false" and the OverflowToolbar works as it has
worked before.

However, when it is set to "true", the _doLayout function of
sap.m.OverflowToolbar is called with delay with requestAnimationFrame.
Thus, it is not blocking the thread immediately after onAfterRendering
and handleResize functions. Also, when it is executed at the very
beginning of the new frame of the Browser, all DOM measurements are
done in the same frame (in the best case scenario) and trigger just
one more styles recalculation of the Browser, leading to performance
improvements. Also, frame drops are less with the new implementation,
which leads to smoother experience upon resizing.

Change-Id: I1ca90ddfb1d88cfb5daf578d3826e8b66a9b3ff1
JIRA: BGSOFUIPIRIN-3925