[FIX] Form: Empty Link is too small
if a link is empty it consumes no space. This leads to a flickering if
the text comes later. Even if it is empty it looks stange.

For Text in display mode form the size is too small if the label is
above the text or no label is there. 

Change-Id: Ie8d312ab30a16f1d5cd150114bed6f27862450b3
BCP: 1980241510