[FIX] sap.m.Select: Fixed error upon scrolling when Select is opened
Problem: When sap.m.Select is opened and the page is getting scrolled,
a console error is thrown because the maximum call stack is reached.

Solution: Preventing the infinite opening/closing of the sap.m.Select
upon scrolling with introducing a new boolean variable, which indicates
when the control is opened/closed.

Change-Id: I23733a3bae794979a92b233632b296b35a673d38
BCP: 1980277274