[FIX] Table: Vertical Scrollbar position
Calculation of the firstVisibleRow can be inaccurate in case of large
data or when zoomed in Chrome. In this case it cannot be scrolled to the
last row. To overcome this issue we consider the table to be scrolled to
the end when the scroll position is less then 1 pixel away from the
maximum

Change-Id: I8ecda6de74608ba75a75ca03b00fe5d641302cb8
BCP: 0020751294 0000270038 2019
CR-Id: 002075125900003495922019