[FIX] sap.ui.unified.ColorPicker: Now the width of the inputs is correct
The ColorPicker is a composite control which consists of a lot of
controls in a small space so every pixel matters. That's why according
to the VD spec, it was needed to override the input's min-width.
As a result of that, the visual tests are now fixed.

BCP: N/A

Change-Id: Ia14419f86259a184d2a197e42600a1e21e9cde2f