[FIX] sap.m.OverflowToolbar: ToolbarSpacer width measurement is fixed
Problem: When ToolbarSpacer has specified width, it is not taken into
account while OverflowToolbar is making calculations to decide whether
there is enough space for all its content or not. As a result, in some
edge cases, the content does not fit but the overflow button is not
shown (because of wrong calculations) and part of the content gets cut.

Solution: The width of the ToolbarSpacer is taken into account when it
is specified with the width property. When not specified, it is not
taken into account for the calculations because it is flexible and
depends on the available space.

Change-Id: Ib03324140484b40dfabffb9a5bc1188841284980
BCP: 1980270564