[FIX] Theming Parameters: Fix urls when UI5 is loaded from other domain
Resolving the urls was done correctly, but only the "path" segment of
the absolute URL has been used. So when UI5 is loaded from a different
domain, this does not work and the url points to the domain the
document is loaded from, not where the UI5 theme resources are located.

This fixes a regression caused by
4576c6899b34f81fb5e5d107eb2da9806be1b0ed

Also fixes a wrong path for the stylesheet test case.

Change-Id: I348391fc087c2a3e06f5b2708c8cf928a677ffab
BCP: 1970102993