[FIX] sap.ui.dt - designtime add or insert plugin handling
On addPlugin, insertPlugin the plugin is added to the designtime but
also the designtime gets added to the plugin. The order of fire addPlugin
event and setDesigntime has changed. addPlugin is fired first, it is
triggers the attach getBusy events to the plugin. After that setDesigntime
is called. For notfication when the designtime is finished with addPlugin
you need to attach to the sync event.

The onBusyChange method is changed. Previously it memorized the added
taskid in a closure. The taskid is needed to close the task.
With this solution it was dificult to add missing detach functions.
The solution now is to close the tasks by type and plugin name. Now
it is not necessary to memorize the taskid anymore.

Change-Id: I18b480f26620c0b4ff45fcfa41d7f3d27e9eda0c
Signed-off-by: Arthur Trauter <arthur.trauter@sap.com>