[FIX] XMLComposite: Only load resource model when needed
Note:
Customer complains that by default the XMLComposite load unavailable
i18n messages bundles. By default the composite loads the i18n message
bundle of the library, even if this is not used.
With this change we check wheter the the resource bundle is used in the
XML representation of the composite.

Note 2:
With that change, whenever there comes a new XML content to the
composite is a UI5 control tree this approach does not work out.
Nevertheless such an approach is not supported in openui5 so far.

Change-Id: Ia65de231bd139af1c6f8a4d0a873f09600a4803f
BCP: 002075129400002643072019