[INTERNAL] Generic tests: avoid issues with require.js
Generic tests load all libraries contained in the current build setup.
When executed in the SAPUI5 layer, this includes sap.viz. This may cause
issues with third party software (like signals.js) which reacts on the
presence of require.js with a different behavior (according to UMD
pattern).

As long as the loader does not provide a mean to prevent this in async
execution mode, the generic tests simply replace require.js by an empty
dummy to avoid the undesirable side effects.

Change-Id: Ib93180bff6cc62480e4995c39e5ec86630305dc0