[FIX] sap.m.SearchField: Improve handling of focus on iOS
When the user clicks in the place of the reset button,
while the reset button is not visible, the input has to be focused.
This was not working well on iOS - the keyboard did
appear only for short time.
Now it is fixed by attaching to the click event of the form.

Snippix: #51088

Change-Id: I45597aa7a14e2dde32ecc08fc673dc6c3c0fe1ab
BCP: 002075129500002240182019