[FIX] sap.m.SplitContainer: Show detail page on phone
When split container is opened on a phone,
the intial page should be the detail page.

NavContainer does not support setInitialPage in
a late state. So now we make sure that it is set as eary as possible.

Change-Id: Ie3b6bc779f04228a394d4f7cfcd1433d9b17fd00
BCP: 1980297329
BCP: 002028376500005408012018
(cherry picked from commit 74e3f7cccb318bbc2e56fa093a8d566b2488d89e)
CR-Id: 002075125800002375032019