[FIX] sap.m.OverflowToolbar: GenericTag is not truncated now
Problem: When GenerticTag, placed in OverflowToolbar in the title of a
DynamicPageTitle, receives its data with a delay, its content gets
truncated.

Solution: When OverflowToolbar dispatches event for content size change,
it now sends additional parameter "invalidate", which is "true" when the
OverflowToolbar has flexible content (such as GenericTag, which
implements new interface). The DynamicPageTitle gets invalidated when
this property is set to "true", so that it can correctly updates
flex-basis property of its content.

Change-Id: Icd1b2b0076a8d4ff3bda8501c2fc850c22142915
BCP: 1980241288