[INTERNAL][FIX] Table: Avoid unnecessary second request in Auto mode
If the table's binding is refreshed and the table's sizes change at the
same time in visibleRowCountMode=Auto, it could happen that the table
triggers 2 requests, because the length isn't calculated correctly.
This change ensures that the length is also calculated correctly for the
first call to getContexts after a binding refresh.

Change-Id: Ie4e8cf8003940ec91cde13044724676db17402fd
BCP: 1970197873