[INTERNAL][FIX] sap.m.SinglePlanningCalendar: Appointment selected state is now read
    The appointment HTML element has 'role' attribute with value 'gridcell' and
all appointments are in HTML element container with 'role' attribute and value 'grid'.
With that precondition when an appointment is focused its selected state is read
on 'JAWS2019' with google 'Chrome' browser, no matter if the appointment 'selected'
state is 'true' or 'false'.
    The issue is now fixed by adding an 'aria-selected' attribute to each
appointment HTML element. The 'aria-selected' attribute value is toggled
when 'appointmentSelect' event is fired.

Change-Id: I3cfc3bda2c73cb79a5756c1aced58bb10bb0adfa
BCP: 1980099454