[INTERNAL] sap.m.Input: showSuggestions setter fixed
Issue:
Setting showSuggestion on/off  multiple times
breaks the suggestion popover and it no longer appears.
Also in cases where the startSuggestion property is "0"
the suggest was triggered when showSuggestion was "false".

Solution:
When focusing the input we check if showSuggestion is true
before triggering suggest.
When setting showSuggestion to true the suggestion popover
popup will be created if needed.

BCP: 1980265707
Change-Id: I455df9bd9c523dba931594e6cda4adeff011f6e1