[INTERNAL][FIX] sap.f.ShellBar: Fixed Incorrect OverflowToolbar layout
Problem: When we have a scenario where the OTB would be with different
background compared to the div of the ShellBar, there are
empty (transparent) spaces on both sides of the ShellBar (its paddings).

Solution: We remove the paddings from the ShellBar and put them on the
OTB. This way we have a consistent background of the whole control at
any use case.

Change-Id: Ic6adb38ab8d7aea909247ce08070d124de73271a
BCP: 1970216613