[INTERNAL] StandardListItem: fix style rule
Fixing a style rule that was caused rendering issue in IE when the
StandardListItem is placed in a Dialog. The correctly flex style now
added to the addressed issue.
GrowingEnablement reused the style classes from the StandardListItem
which caused rendering issue of the trigger button. The adaptation of
the correct style classes is not added for fixing the issue.
BCP: 1970217362

Change-Id: I41f9f98bb72048949ea90bf00bab4bf22f88dad6