[FIX] sap.ui.fl.FlexController:
When changes from different controls are processed after controls are
re-created.

When the a change is successfull it still shows as a dependency for
other incoming changes due to a failed check (_checkIfDependencyIsStillValid).
Also an unprocessed change by _applyChangesOnControl
could  already added to mDependentChangesOnMe leading to an error in
_updateDependencies().

Change-Id: I4f4cba58eae121fe77df85802e05054c0093a8f4
BCP: 002075129400002782942019