[FIX] sap.m.UploadCollection : Download of item not happening when factory binding is used.
Problem:
The item is not getting downloaded when fileName is clicked and in console dump is seen under triggerLink method.
This happens only when factory binding is used in items aggregation.

Fix:
Earlier the fileName id was used to determine the item position.FileName generated differs based on template or factory method used.
Hence it was not reliable.Added fix so that the download url is taken from item directly.

Change-Id: I6110af3cfb0b67f130d4f0f650c492e1416dca73
BCP: 012005041000000451312019.