[FIX] sap.f.DynamicPage: Improved Runtime Adaptation (RTA) behavior
Problem: When the DynamicPageHeader is "pinned" or the property
"preserveHeaderStateOnScroll" of the DynamicPage is true, we can't
manipulate the elements in the DynamicPageHeader while under RTA.

Root Cause: The configuration for the "scrollContainers" which is
important for the RTA logic was incorrect.

Solution: We set the correct "scrollContainers" configuration when the
DynamicPageHeader is in pinned state or when the
preserveHeaderStateOnScroll property of DynamicPage is set to true.

Change-Id: I3e9f15cd4eae79fc79f062bda8691795cc617f68
BCP: 1970246385