[FIX] sap.uxap.ObjectPageLayout: Improved Runtime Adaptation (RTA) behavior
Problem: When the ObjectPageDynamicHeaderContent is "pinned", or the
property "preserveHeaderStateOnScroll" of the ObjectPageLayout is true,
when can't manipulate the elements in the ObjectPageDynamicHeaderContent
while under RTA.

Root Cause: The configuration for the "scrollContainers" which is
important for the RTA logic was incorrect.

Solution: We set the correct "scrollContainers" configuration in the
case of "pinned" ObjectPageDynamicHeaderContent or
"preserveHeaderStateOnScroll" property of the ObjectPageLayout set to
true.

Change-Id: I7b5f269fb9b6b8ca74f1fa748fa475e5bf300f19
BCP: 1970246385