[FIX] sap.f.SemanticPage: Improved Runtime Adaptation (RTA) behavior
Problem: When the SemanticPageHeader is "pinned", or the
property "preserveHeaderStateOnScroll" of the Semantic Page is true,
we can't manipulate the elements in the SemanticPageHeader
while under RTA.

Root Cause: The configuration for the "scrollContainers" which is
important for the RTA logic was incorrect.

Solution: We set the correct "scrollContainers" configuration when the
SemanticPageHeader is in pinned state or when the
preserveHeaderStateOnScroll property of SemanticPage is set to true.

Change-Id: I87d8566c8ba3913c5317a5dd9496151d3a06a017
BCP: 1970246385