[INTERNAL] Gherkin: further reduce need for 'eval'
Before this change, integration tests in demokit samples still have used
'eval' in some cases. This was caused by the pattern for eligible step
names, which didn't accept digits in page or method names.

Furthermore, the component used as 'app under test' in the
GherkinWithPageObjects sample didn't use a manifest yet which caused
sync loading of the controls of its root view.

Change-Id: Ifa954d3692d3156230fdca461be171843828fcad