[INTERNAL] ListBase: context menu fix
When setting the binding context of the context menu the model name was
not passed and due to this when the table is bound to a named model, the
context menu did not give the expected result. To fix this issue, we now
pass the model name when setting the binding context to the context menu.
The API documentation is also improved for the same.
BCP: 1970253997

Change-Id: Ia6a8d0802b29cb6c6824a0d75747e44d827b5b9b