[FIX] performance/trace/FESR: Beacon strategy adaptations
Passport Action:
As the count may get lost if the string length exceeds the
field length (40 characters), it is necessary to perform
the trimming from the left side.

FESR Transaction ID:
When using the beacon strategy, the Transaction ID should
be added also for UI-only FESR.

Visibility Change:
As the unload event is synchronous this should be avoided.

Beacon Body:
Minor adaptation to the Beacon body formatting.

Beacon Timeout:
Beacon is sent after 60s, if no other threshold is met.

Change-Id: Ifa7cf1a2458b1bc5404abe714a858b55ca1dfa37