[INTERNAL] sap.uxap.ObjectPageSubSection: Ensured height includes padding
Background:
ObjectPage subSection with class sapUxAPObjectPageSubSectionFitContainer
are internally given specific height (equal to the height of the sections
container)

Problem:
If additional CSS overstyles the subSection to give it padding
=> the final subSection height is not the expected one
(in this case smart templates have overstyling CSS)

Solution:
Ensure the height (that objectPage sets to its subSections) is not
affected by changes in the paddings.
The solution was already applied for fiori3, but now we see that
it is required for any theme.

Change-Id: I51086ed421d627efb4e97433862de0940b7c4af1
BCP: 1970113301