[FIX] sap.m.IconTabBar: Ensure only one tab is selected at a time
When the content of a tab was updated through data binding
updateAggregation was called, which nullified the current selectded item.
But if the binding is done through a template, no invalidation happens.
Then when switching to a new tab, we could have 2 selected tabs at
the same time.

Solution: Ensure the header is re-rendered always when updateAggregation
is called.

Change-Id: I806dcb5dfc921722234d708d17ec5e0a373ff80f
Snippix: 50906
BCP: 1970197493