[FIX] sap.m.Token: ARIA description updated based on parent
- In order for the ARIA attributes to be updated without
introducing backward incompatibility, a new hidden boolean
property was added to the sap.m.Token control.
- When this property is set to true, the text of the
'deletable describe by' property is added to the token,
and when set to false - it is removed.
- This is needed, as there might be cases when some of the
tokens are editable and others are not. If we overwrite the
editable property itself, we will change the state of the
tokens, which may lead to incompatibility.

Change-Id: I3df3e62d534cee45d72b1da8e86a7ebab56840ca
BCP: 1970272491