[FIX] sap.m.OverflowToolbar: Empty Popover is not shown
Problem: When "visible" property of some of OverflowToolbar's buttons
gets changed between the execution of init and doLayout functions of
OverflowToolbar, the button still stays in the movable controls
collection. That's why it goes in the Popover, but as it is not visible,
when opened, the Popover looks empty.

Solution: Update all controls collections of OverflowToolbar on property
change, no matter if we listen for changes at that point or not.

Change-Id: I4ee1ee231ae275631a36e731e9b00686da3238f8
BCP: 1970266072