[INTERNAL] StandardListItem: flex flex behavior
As in the Visual Spec, when there is title and info text available, they
should share 50% of the width. But this spec was never implemented and
was introduced in the 1.67 version with the renovation of the
StandardListItem control. The 50% flex behavior causes regression in the
ComboBox control as the additionText is truncated even when there is
space available. To be compatible to the old behavior of the
StandardListItem, the 50% flex behavior is removed. Perhaps, a property
can be provided to enable the above 50% flex behavior for compatibility
reasons.
BCP: 1970263556
BCP: 1970310844

Change-Id: Idf9ffdbb1a80f99c0c55a9fed6bfebfbec0dc9fd