[INTERNAL] sap.m: postpone tests until theme has been applied
Recent test executions in CI show a higher rate of test failures and
retries due to incomplete or missing rendering. This might be caused by
the fact that test execution now uses browser caching.

To avoid  the overhead of the retries, all mobile tests now use
"xx-waitForTheme:init" to postpone the test until after the theme has
been applied. Only the "Rule" test starts earlier as the RuleSet engine
seems to have timing issues when its start is postponed. This needs to
be fixed separately.

Change-Id: Idd0a3f503e6e355a13ae14cc1922a63882a35177