[FIX] OverflowToolbar: Resize is now handled correctly when in Dialog
Problem: If the "OverflowToolbar" is in "Dialog", on iPhone 6/7/8
there is a bug in the layout when the orientation is changed
from portrait to landscape.

Root Cause: "doLayout" cache small values of the content, which results
to wrong calculations when the width of the "OverflowToolbar" gets
bigger
when the orientation changes from portrait to landscape.

Solution: We reset and rerender the "OverflowToolbar" with the newly
added
property "_bForceRerenderOnResize" when needed. In our case in the
Dialog.

Change-Id: Ib566c9a38a32d839c143e47846c4c5bc658eea85
BCP: 002075129400002140952019
CR-Id: 002075125800003844182019