[INTERNAL] adding the sap.ui.fl.apply.internal.Connector
This change adds the Connector.js as well as the basis functionality to
require the default connector
(sap.ui.fl.apply.internal.connectors.LrepConnector)
as well as loads the connector for the changes bundle
(sap.ui.fl.apply.internal.connectors.StaticFileConnector).

In addition the sap.ui.core.Configuration was adjusted to detect this
new connectors and still works with the disabling of an flex request
(set the flexibilityServices parameter empty).

Change-Id: Id0d0c9c8acef1ac26ece5296721c6128361c16c1
JIRA: UICSFLEX-5447