[INTERNAL] adding sap.ui.fl.write.internal.Connector
This change adds the Connector.js as well as the basis functionality to
require the default connector
(sap.ui.fl.write.internal.connectors.LrepConnector)
and loads the connector for the changes bundle
(sap.ui.fl.write.internal.connectors.StaticFileConnector).

In addition the sap.ui.core.Configuration was adjusted by removing the
URL and storing it directly in the corresponding connectors.

Change-Id: I1f5b1fc53d4b9e6fb42a4fc243ba8257fdde5f1b
JIRA: UICSFLEX-5447