[FIX] sap.m.InputBase: Icon size fixed
1px border is added to the icon on hover and in active state.
This caused miscalculation in the input width, leading to pushing
the icon to a second row inside the layout of the input. As a result,
the icon was flickering on hover and press in High Contrast Black theme.

The icon size is now preserved by adding a margin.

Change-Id: Id3aded18f623fe1f58199b56199e55604c864532
BCP: 1970336083
BCP: 1970378775
BCP: 0020751295 0000413798 2019
CR-Id: 002075125800004356842019