[FIX] sap.m.OverflowToolbar: Srinkable Controls overflow correctly
Problem: Controls with shrinkable LayoutData and min-width, specified in
rems or percents do not overflow correctly.

Solution: When shrinkable LayoutData with specified min-width is applied
to a Control, the min-width is extracted from the DOM reference, if it
is not relative, and if it is relative - it gets calculated, taking into
account the width of the OverflowToolbar itself.

Change-Id: I4689dd1b70b3e941eab547defc071f482eec63ec
BCP: 0020751295 0000258142 2019