[FIX] sap.m.IconTabHeader: Tab labels reading by JAWS is imporoved
If focus is moved outside the window and then returned again in the window,
all tab labels were read out by JAWS 2019.
If 'role=tablist' is set to the direct parent of the tabs, this is fixed.
Also accessibility information in Chrome Dev Tools is correct with this change.

Change-Id: I76883d1028197b89a9f05c3b05caac4911bab53b
BCP: 1970191138